refactor(frontend): SkyBackground を責務ごとに分割し図形を外部化
- 単一ファイルだった SkyBackground を components/sky/ 配下に分割
  - skyConfig: 雲・飛行船の配置データと型
  - skyGeometry: 雲の静的ジオメトリ（楕円座標）の定数定義
  - airship.svg: 飛行船の図形を SVG ファイルに分離（SVGR で取り込み）
  - CloudShape / Airship: SVG 描画（色・seed・ID など動的値の付与に専念）
  - SkyBackground: オーケストレーター
- @svgr/webpack を導入し、next.config に dev(turbopack)/build(webpack) の loader を設定
- 飛行船の進行方向を Record で扱い、方向ごとの指定を対称に記述
- 冗長なコメント・他レイヤーの責務に触れるコメントを除去
