+++ frontend/package.json
@@ -34,6 +34,7 @@
     "unist-util-visit": "^5.1.0"
   },
   "devDependencies": {
+    "@svgr/webpack": "^8.1.0",
     "@tailwindcss/postcss": "^4",
     "@types/mdast": "^4.0.4",
     "@types/node": "^20",
