+++ frontend/src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
-import SkyBackground from "./components/SkyBackground";
+import SkyBackground from "./components/sky/SkyBackground";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
